Single Store!
Yay! We call our array only once. Hurray.
